# KisaanCafe.Models/ProductCommand.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace KisaanCafe.Models
{
    public class ProductCommand
    {
        public int Id { get; set; }
        public string Name { get; set; }
        public string Description { get; set; }
        public decimal Prize { get; set; }
        public string size { get; set; }
        public byte[] ImageData { get; set; }
    }
}