# KisaanCafeWebAPI/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "KisaanCafeWebAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Product/GetAllProducts": {
      "get": {
        "tags": [
          "Product"
        ],
        "operationId": "GetAllProducts",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Product/AddProduct": {
      "post": {
        "tags": [
          "Product"
        ],
        "operationId": "AddProduct",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Product/AddInvoice": {
      "post": {
        "tags": [
          "Product"
        ],
        "operationId": "AddInvoice",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/InvoiceDetails"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Product/{productId}": {
      "put": {
        "tags": [
          "Product"
        ],
        "operationId": "UpdateProduct",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/PutActionResult"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "operationId": "DeleteProduct",
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "InvoiceDetails": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "fullName": {
          "type": "string"
        },
        "fullAddress": {
          "type": "string"
        },
        "paymentMode": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "totalAmount": {
          "format": "int64",
          "type": "integer"
        },
        "ammountPaid": {
          "format": "int64",
          "type": "integer"
        },
        "amountLeft": {
          "format": "int64",
          "type": "integer"
        },
        "invoiceProductDetails": {
          "$ref": "#/definitions/InvoiceProductDetails"
        }
      }
    },
    "InvoiceProductDetails": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "prize": {
          "format": "double",
          "type": "number"
        },
        "weight": {
          "format": "double",
          "type": "number"
        },
        "totalWeight": {
          "format": "double",
          "type": "number"
        },
        "totalPrize": {
          "format": "double",
          "type": "number"
        },
        "invoiceDetailsId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "ProductCommand": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "prize": {
          "format": "double",
          "type": "number"
        },
        "weight": {
          "format": "double",
          "type": "number"
        },
        "imageData": {
          "format": "byte",
          "type": "string"
        }
      }
    },
    "PutActionResult": {
      "type": "object",
      "properties": {
        "newResourceRelativeUrl": {
          "type": "string"
        },
        "result": {
          "$ref": "#/definitions/PutActionResultCode"
        },
        "newResourceValue": { }
      }
    },
    "PutActionResultCode": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "type": "integer"
    }
  }
}